+++ src/connection/base.js
@@ -35,7 +35,8 @@ class ConnectionBase {
   _dataChannels: Array<window.RTCDataChannel>;
   _callbacks: Object;
   _pcConfig: {
-    iceServers: Array<Object>
+    iceServers: Array<Object>,
+    iceTransportPolicy: string,
   };
 
   /**
@@ -46,7 +47,7 @@ class ConnectionBase {
       this._callbacks[kind] = callback;
     }
   }
-  constructor(signalingUrl: string, roomId: string, options: ConnectionOptions, debug: boolean = false) {
+  constructor(signalingUrl: string, roomId: string, options: ConnectionOptions, debug: boolean = false, isRelay = false) {
     this.debug = debug;
     this.roomId = roomId;
     this.signalingUrl = signalingUrl;
@@ -61,7 +62,8 @@ class ConnectionBase {
     this._isOffer = false;
     this.connectionState = 'new';
     this._pcConfig = {
-      iceServers: this.options.iceServers
+      iceServers: this.options.iceServers,
+      iceTransportPolicy: isRelay ? "relay" : "all"
     };
     this._callbacks = {
       open: () => {},
@@ -163,6 +165,7 @@ class ConnectionBase {
   }
 
   _createPeerConnection() {
+    this._traceLog('RTCConfiguration=>', this._pcConfig);
     const pc = new window.RTCPeerConnection(this._pcConfig);
     const audioTrack = this.stream && this.stream.getAudioTracks()[0];
     if (audioTrack && this.options.audio.direction !== 'recvonly') {